Fix some of the possibles failure reasons in e2e tests (#294)
